feat(admin-ui): draw the campaign journey as a flow, not a bar
Feedback, and it was right: the campaign screen is unusable for the people it is
for. A CDP draws a journey as a canvas — entry, steps, branches — and we drew a
grey strip wedged between tables of UUIDs, with `SUPPRESSED_QUIET_HOURS` still
the vocabulary.

What a campaign person needs, in order: how many entered, what happened at each
step as a path read left to right, and where the ones who dropped out went and
why. A flow answers all three at a glance because it has the shape of the thing
it describes; a bar answers the second only, and only after a legend.

Drawn as SVG following docs/service-map — the console already has a hand-drawn
canvas for this job, and a third visual language would be worse than either.
Colour stays on the ADR-0208 D2 variables, never a literal.

Two defects found by rendering it and looking, not by tests:

- the last node's drop labels rendered past the viewBox, so a count was simply
  not on screen — data hidden by layout, which is worse than ugly;
- three branches shared one origin and drew through the same pixels, reading as
  one thick line rather than three outcomes.

The raw outcome moved to a data attribute. It was an SVG <title> inside <text>,
which counts as text content — so the enum was "on the screen" for anything
reading textContent, including the guard that exists to keep enums off it.

Also splits `app/api/infra/status/route.ts`: it exported probe helpers so a test
could import them, which Next forbids in a route file. Turbopack tolerates it,
webpack rejects it outright, and it was the single blocker to building with the
only bundler that emits client source maps (#3235). Probe logic a test drives
directly does not belong in a request handler anyway.

Refs #2895, #3235.
