+++ openbank-admin-ui/src/test/journey-flow.test.tsx
@@ -6,12 +6,12 @@ import React from 'react'
 import { describe, expect, it } from 'vitest'
 import { render, screen } from '@testing-library/react'
 import { LanguageProvider } from '@/lib/i18n/LanguageContext'
-import { JourneyFlow } from '@/components/campaigns/JourneyFlow'
+import { JourneyCanvas } from '@/components/campaigns/JourneyCanvas'
 
 /**
- * The journey view exists because the send-log table is unreadable for the person who decides
- * whether a campaign is working. These pin the three properties that make it readable — each of
- * which is easy to lose in a later edit precisely because the table breaks all three on purpose.
+ * The journey is drawn as a flow because the send-log table is unreadable for the person deciding
+ * whether a campaign is working. These pin the properties that make it readable — each easy to lose
+ * in a later edit, and each one the table breaks on purpose.
  */
 
 const STEPS = [
@@ -37,85 +37,73 @@ function renderFlow(funnel = FUNNEL, audience: number | null = 1000) {
     React.createElement(
       LanguageProvider,
       null,
-      React.createElement(JourneyFlow, { steps: STEPS, funnel, audienceSize: audience }),
+      React.createElement(JourneyCanvas, { steps: STEPS, funnel, audienceSize: audience }),
     ),
   )
 }
 
-describe('journey flow', () => {
+describe('journey canvas', () => {
   /**
-   * The whole point of the screen. `SUPPRESSED_QUIET_HOURS` is a value the API returns; a marketer
-   * needs "it was night". If an enum leaks into the visible text, this view has become the send-log
-   * table with extra steps.
+   * The whole point. `SUPPRESSED_QUIET_HOURS` is a value the API returns; a marketer needs
+   * "quiet hours". If an enum reaches the visible text this has become the send-log table with
+   * extra steps.
    */
   it('never puts a raw enum on the screen', () => {
     const { container } = renderFlow()
 
     expect(container.textContent).not.toMatch(/SUPPRESSED_/)
     expect(container.textContent).not.toMatch(/\bFAILED\b/)
-    expect(screen.getByText(/No marketing consent/)).toBeTruthy()
+    expect(screen.getByText(/No consent/)).toBeTruthy()
     expect(screen.getByText(/Quiet hours/)).toBeTruthy()
     expect(screen.getByText(/Delivery failed/)).toBeTruthy()
   })
 
   /**
    * …but the raw value must stay reachable, or the screen and the API can drift apart with nobody
-   * noticing. Relabelling is a translation, not a replacement.
+   * noticing. It lives in a data attribute rather than an SVG <title>, because a <title> inside
+   * <text> IS text content and would defeat the assertion above.
    */
   it('keeps the raw outcome reachable for whoever is debugging', () => {
-    renderFlow()
+    const { container } = renderFlow()
 
-    expect(screen.getAllByTitle('SUPPRESSED_CONSENT').length).toBeGreaterThan(0)
-    expect(screen.getAllByTitle('SUPPRESSED_QUIET_HOURS').length).toBeGreaterThan(0)
-    expect(screen.getAllByTitle('FAILED').length).toBeGreaterThan(0)
+    expect(container.querySelector('[data-outcome="SUPPRESSED_CONSENT"]')).toBeTruthy()
+    expect(container.querySelector('[data-outcome="SUPPRESSED_QUIET_HOURS"]')).toBeTruthy()
+    expect(container.querySelector('[data-outcome="FAILED"]')).toBeTruthy()
   })
 
-  it('states the delivered share against what the step reached', () => {
+  it('states what entered, what was delivered, and what each branch lost', () => {
     renderFlow()
 
-    expect(screen.getByText(/600/)).toBeTruthy()
-    expect(screen.getByText(/delivered of/)).toBeTruthy()
-    expect(screen.getByText(/60 %/)).toBeTruthy()
-  })
-
-  /**
-   * A later step still waiting out its delay has reached nobody. Rendering that as "0 delivered of
-   * 0" puts three numbers on screen that look like a result and are the absence of one — and on a
-   * multi-step journey that is most of the screen.
-   */
-  it('says a step has reached nobody rather than printing zeroes', () => {
-    renderFlow()
-
-    expect(screen.getByText(/nobody yet/)).toBeTruthy()
-    // …and exactly one step reports a real result, so the zero row cannot be mistaken for one.
-    expect(screen.getAllByText(/delivered of/)).toHaveLength(1)
+    expect(screen.getAllByText('1,000').length).toBeGreaterThan(0)
+    expect(screen.getByText('600')).toBeTruthy()
+    expect(screen.getByText('300')).toBeTruthy()
   })
 
   /**
-   * A campaign that has not run yet must say so. An empty funnel rendered as bars reads as "nobody
-   * matched", which is a business conclusion, and the opposite of the truth.
+   * A campaign that has not run yet must say so. An empty flow drawn with zeroes reads as "nobody
+   * matched", which is a business conclusion and the opposite of the truth.
    */
-  it('says nothing has run rather than drawing an empty funnel', () => {
+  it('says nothing has entered rather than drawing zeroes', () => {
     renderFlow([], null)
 
-    expect(screen.getByText(/has not reached anyone yet/)).toBeTruthy()
+    expect(screen.getByText(/Nobody has entered yet/)).toBeTruthy()
   })
 
   /** The wait between steps is part of the design; seconds are not a unit anyone reads. */
   it('renders the delay between steps in human units', () => {
     renderFlow()
 
-    expect(screen.getByText(/after 2 d/)).toBeTruthy()
+    expect(screen.getByText(/wait 2 d/)).toBeTruthy()
     expect(screen.queryByText(/172800/)).toBeNull()
   })
 
   /**
-   * The audience number is stated as pre-consent, pre-suppression. Without that it reads as the
-   * number of people who got the campaign, which is the single most expensive misreading here.
+   * The entry number is the segment size, stated as such. Without that it reads as the number of
+   * people who got the campaign — the single most expensive misreading on this screen.
    */
-  it('qualifies the audience number so it cannot be read as reach', () => {
+  it('labels the entry number as a segment, not as reach', () => {
     renderFlow()
 
-    expect(screen.getByText(/before consent checks and suppression/)).toBeTruthy()
+    expect(screen.getByText(/people in segment/)).toBeTruthy()
   })
 })
